Recompile contracts before deployment
